backend/pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "fast-api-docker-poetry"
version = "0.1.0"
description = "Fast Api voice chat bot API" 
authors = ["xsnyrch"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.95.1"
uvicorn = "^0.18.2"
alembic = "^1.10.4"
python-multipart = "^0.0.5"
python-ulid = "^1.1.0"
backoff = "^2.2.1"
fastapi-restful = "^0.4.3"
sqlalchemy = "^2.0.12"
asyncpg = "^0.27.0"
psycopg2-binary = "^2.9.6"
bcrypt = "^4.0.1"
passlib = "^1.7.4"
python-jose = { extras = ["cryptography"], version = "^3.3.0" }
opentelemetry-distro = "^0.39b0"
opentelemetry-exporter-otlp = "^1.18.0"
opentelemetry-contrib-instrumentations = "^0.39b0"
piper-tts = "^1.2.0"
lmstudio = "^1.3.0"

[tool.poetry.group.dev.dependencies]
black = "^22.6.0"
pyproject-flake8 = "^0.0.1a4"
pre-commit = "^2.19.0"
pytest = "^7.1.2"
pytest-cov = "^3.0.0"
types-PyYAML = "^6.0.9"
pytest-asyncio = "^0.21.0"
ipykernel = "^6.29.5"

[tool.black]
line-length = 100
target-version = ['py310']

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
log_cli = true
log_cli_level = "INFO"
log_cli_format = "%(asctime)s [%(levelname)8s] %(message)s (%(filename)s:%(lineno)s)"
log_cli_date_format = "%Y-%m-%d %H:%M:%S"

[tool.pydantic-pycharm-plugin]
parsable-type-highlight = "disable"   # This is defined as parsable so don't warn on it
acceptable-type-highlight = "warning"
